+++ app/Http/Controllers/CategoriaReceberController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\CategoriaReceber;
+use App\Models\ContaReceber;
+
 
 
 class CategoriaReceberController extends Controller
@@ -17,6 +19,12 @@ class CategoriaReceberController extends Controller
 
     //CADASTRO DE CATEGORIA DE CONTAS A RECEBER
     function cadastrar($usuario, Request $request){
+
+        //Validar para não salvar descrição
+        $validated = $request->validate([
+            'descricao' => 'required',
+        ]);
+
         //Definindo data para cadastrar
         date_default_timezone_set('America/Cuiaba');    
 
@@ -31,6 +39,17 @@ class CategoriaReceberController extends Controller
     //EXCLUIR DE CONTAS A PAGAR
     function excluir($id){
         $categoria_receber = CategoriaReceber::find($id);
+
+         //Verifica se existe
+         if (!$categoria_receber) {
+            return redirect()->back()->with('error', 'Descrição não encontrada.');
+        }
+
+        // Verifique se existem relacionamentos
+        $relacionamentos = ContaReceber::where('categoria_receber_id', $categoria_receber->id)->get();
+        if ($relacionamentos->count() > 0) {
+            return redirect()->back()->with('error', 'Esta descrição está relacionado a alguma conta a receber e não pode ser excluído.');
+        }
         $categoria_receber->delete();
         return redirect()->back()->with('success', 'Exclusão realizada com sucesso');
